Scope introduced
Mostly straightforward, C-style scope with blocks adding a new level of
scope. Once functions are introduced, will have to keep track of
function frames, too, so that we can just store variables in this same
scope structure but skip other function scopes. We wouldn't want the set
of accessible variables to be dependent on the call stack because that's
chaos. But having a single, global scope stack is nifty and versatile.