Start to type declarations and checking
Variables are now declared with a type and on assignment, if the type
is incompatible, it fails. Currently there's only checking at runtime
using the actual value of the variable. Also, there's no inheritance
yet, so it's a straight check of the type.

Currently type declarations are mandatory, but in the future I'll remove
this to allow to type inference (we currently can do that at runtime, of
course, but until the type declarations are everywhere, we cannot
statically check the program).