Static type checking working for builtins
Function signatures are internally stored like
`Type$method__ArgType1__ArgType2`. Eg, `Int$__add__Int` would be used
for the `__add` method on type `Int` that takes in an `Int`. This name
mangling is done for uniqueness and should be an easy optimization to
make in the future (to replace calls to, eg, `__add` with `__add__Int`,
etc).